move all space related defaults files under space.defaults module
