+++ src/domain/challenge/challenge/challenge.service.ts
@@ -38,8 +38,8 @@ import { AgentService } from '@domain/agent/agent/agent.service';
 import { AgentInfo } from '@src/core/authentication/agent-info';
 import { limitAndShuffle } from '@common/utils/limitAndShuffle';
 import { CommunityRole } from '@common/enums/community.role';
-import { challengeCommunityPolicy } from './challenge.community.policy';
-import { challengeCommunityApplicationForm } from './challenge.community.application.form';
+import { challengeCommunityPolicy } from '../space.defaults/definitions/challenge.community.policy';
+import { challengeCommunityApplicationForm } from '../space.defaults/definitions/challenge.community.application.form';
 import { ICollaboration } from '@domain/collaboration/collaboration/collaboration.interface';
 import { NamingService } from '@services/infrastructure/naming/naming.service';
 import { LimitAndShuffleIdsQueryArgs } from '@domain/common/query-args/limit-and-shuffle.ids.query.args';
@@ -48,7 +48,7 @@ import { IProfile } from '@domain/common/profile/profile.interface';
 import { OperationNotAllowedException } from '@common/exceptions/operation.not.allowed.exception';
 import { CollaborationService } from '@domain/collaboration/collaboration/collaboration.service';
 import { TagsetReservedName } from '@common/enums/tagset.reserved.name';
-import { challengeDefaultCallouts } from './challenge.default.callouts';
+import { challengeDefaultCallouts } from '../space.defaults/definitions/challenge.default.callouts';
 import { IStorageAggregator } from '@domain/storage/storage-aggregator/storage.aggregator.interface';
 import { SpaceDefaultsService } from '../space.defaults/space.defaults.service';
 import { ISpaceSettings } from '../space.settings/space.settings.interface';
