+++ src/domain/challenge/opportunity/index.ts
@@ -1,5 +1,5 @@
 export * from './opportunity.entity';
 export * from './opportunity.interface';
 export * from './dto';
-export * from './opportunity.community.policy';
-export * from './opportunity.community.application.form';
+export * from '../space.defaults/definitions/opportunity.community.policy';
+export * from '../space.defaults/definitions/opportunity.community.application.form';
