+++ src/domain/challenge/space.defaults/definitions/challenge.community.application.form.ts
