+++ src/domain/challenge/space.defaults/definitions/challenge.community.policy.ts
