+++ src/domain/challenge/space.defaults/definitions/challenge.default.callouts.ts
