+++ src/domain/challenge/space.defaults/definitions/opportunity.community.application.form.ts
