+++ src/domain/challenge/space.defaults/definitions/opportunity.community.policy.ts
