+++ src/domain/challenge/space.defaults/definitions/opportunity.default.callouts.ts
