+++ src/domain/challenge/space.defaults/definitions/space.community.application.form.ts
