+++ src/domain/challenge/space.defaults/definitions/space.community.policy.ts
