+++ src/domain/challenge/space.defaults/definitions/space.default.callouts.ts
