+++ src/domain/challenge/space/index.ts
@@ -1,5 +1,5 @@
 export * from './dto/space.dto.update';
 export * from './dto/space.dto.delete';
 export * from './dto/space.dto.create';
-export * from './space.community.policy';
-export * from './space.community.application.form';
+export * from '../space.defaults/definitions/space.community.policy';
+export * from '../space.defaults/definitions/space.community.application.form';
